Python script to combine docs for google translate (#15)
Idea to combine all the script files together so that I can do google
translate in one shot, then split it up afterwards.

- Possible to do this easily because Chatgpt released after 2022-Nov,
which was after I dropped this project - with chatgpt, it is easier to
write python scripts
- Google translate allows up to 10 MB per file, and total file size of
scripts isn't that huge
- This saves me time instead of dragging each individual script and
downloading each time, lots of clicks

Also found an increase in quality of machine translations, so will be
generating the mtl-script folder again of the remaining scripts
Besides the issue of pronouns you I me he her she him, or 子 translating
to girl instead of boy, due to the vagueness of Japanese language, it
seems very accurate.

Will need to edit pronouns. Will also do a search to find and replace
gender pronouns. But girl sometimes is used in kink scenes like a girl,
so shouldn't mass replace it.

Did I mention that I still have to understand chatgpt's code to fix line
break issue? Ugh. But still saved me a lot of work, thanks chatgpt.